+++ vite.config.ts
@@ -9,4 +9,5 @@ import react from '@vitejs/plugin-react'
 
 export default defineConfig({
   plugins: [react()],
+  base: './',
 })